audio and UI controls desc
