+++ Assets/Audio/sfx/362205__taranp__horn-fail-wahwah-2.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 31abb139c8c3a32469e62e3fb6356dee
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
