+++ Assets/External Packages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97d6eafc92d97644b9d18461751d17c0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
