+++ Assets/External Packages/Music Tracks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3fc53dc05fbd66946a163242df383ed5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
