+++ Assets/Prefabs/Treasure.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 6780543208303325454}
   - component: {fileID: 7390493025879161767}
   - component: {fileID: 848276120508241956}
+  - component: {fileID: -3094187903730578266}
   m_Layer: 0
   m_Name: Treasure
   m_TagString: Treasure
@@ -113,3 +114,25 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 54f160673bfff2345973cc87ef9c4238, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  sfxTreasure: {fileID: 8300000, guid: 4fe0cdca93bdd91409327febb4a888c4, type: 3}
+--- !u!50 &-3094187903730578266
+Rigidbody2D:
+  serializedVersion: 4
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6780543208303325455}
+  m_BodyType: 0
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 0
+  m_Material: {fileID: 0}
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
