+++ Assets/Scripts/EnemySight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64d371ea86bb44c4aaed06abebad6cc6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
