+++ Assets/Scripts/destroyTreasure.cs
@@ -7,7 +7,6 @@ public class destroyTreasure : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        
     }
 
     // Update is called once per frame
